+++ R/crew_client.R
@@ -150,9 +150,11 @@ crew_class_client <- R6::R6Class(
     .started = FALSE,
     .url = NULL,
     .profile = NULL,
-    .condition = NULL,
     .client = NULL, # TODO: remove if/when the dispatcher becomes a thread.
-    .dispatcher = NULL # TODO: remove if/when the dispatcher becomes a thread.
+    .dispatcher = NULL, # TODO: remove if/when the dispatcher becomes a thread.
+    # TODO: remove if/when mirai::status()$events
+    # becomes irrelevant (issue #232):
+    .queue_events = NULL
   ),
   active = list(
     #' @field host See [crew_client()].
@@ -196,10 +198,6 @@ crew_class_client <- R6::R6Class(
     profile = function() {
       .profile
     },
-    #' @field condition Condition variable of the client.
-    condition = function() {
-      .condition
-    },
     #' @field client Process ID of the local process running the client.
     client = function() {
       .client
@@ -241,10 +239,8 @@ crew_class_client <- R6::R6Class(
       .serialization <<- serialization
       .seconds_interval <<- seconds_interval
       .seconds_timeout <<- seconds_timeout
-      # Creating the CV here instead of as a default R6 field value
-      # somehow appeases covr and R CMD check.
-      .condition <<- nanonext::cv()
       .relay <<- relay
+      .queue_events <<- collections::queue()
     },
     #' @description Validate the client.
     #' @return `NULL` (invisibly).
@@ -344,10 +340,10 @@ crew_class_client <- R6::R6Class(
         .dispatcher <<- ps::ps_handle(pid = pid)
       }
       # End dispatcher code.
-      .condition <<- mirai::nextget(x = "cv", .compute = .profile)
-      .relay$set_from(.condition)
+      .relay$set_from(mirai::nextget(x = "cv", .compute = .profile))
       .relay$start()
       .started <<- TRUE
+      .queue_events <<- collections::queue()
       invisible()
     },
     #' @description Stop the mirai client and disconnect from the
@@ -361,7 +357,6 @@ crew_class_client <- R6::R6Class(
         mirai::daemons(n = 0L, .compute = .profile)
       }
       .profile <<- NULL
-      .condition <<- nanonext::cv()
       .relay$terminate()
       .url <<- NULL
       .started <<- FALSE
@@ -396,24 +391,43 @@ crew_class_client <- R6::R6Class(
         error = function(condition) NULL
       )
       # End dispatcher checks.
+      .queue_events <<- collections::queue()
       invisible()
     },
-    #' @description Get the true value of the `nanonext` condition variable.
-    #' @return The value of the `nanonext` condition variable.
-    resolved = function() {
-      nanonext::cv_value(.condition)
-    },
-    #' @description Internal function:
-    #'   return the `mirai` status of the compute profile.
-    #' @details Should only be called by the launcher, never by the user.
-    #'   The returned `events` field changes on every call and must be
-    #'   interpreted by the launcher before it vanishes.
-    #' @return A list with status information.
+    #' @description Get the counters from `mirai::status()`.
+    #' @return A named integer vector of task counts
+    #'   (awaiting, executing, completed) as well as the number of
+    #'   worker connections.
     status = function() {
-      mirai_status(
+      status <- mirai_status(
         profile = .profile,
         seconds_interval = .seconds_interval,
         seconds_timeout = .seconds_timeout
+      )
+      push_events(.subset2(status, "events"))
+      status
+    },
+    #' @description Collect mirai worker-specific
+    #'   connection/disconnection events.
+    #' @return `NULL` (invisibly).
+    #' @param events Integer vector of connection/disconnection events.
+    push_events = function(events) {
+      # TODO: remove when issue #232 is fixed and crew no longer needs
+      # mirai::status()$events.
+      lapply(events, .subset2(.queue_events, "push"))
+      invisible()
+    },
+    #' @description Collect mirai worker-specific
+    #'   connection/disconnection events.
+    #' @return Integer vector of connection/disconnection events.
+    collect_events = function() {
+      # TODO: remove when issue #232 is fixed and crew no longer needs
+      # mirai::status()$events.
+      as.integer(
+        replicate(
+          .subset2(.queue_events, "size")(),
+          .subset2(.queue_events, "pop")()
+        )
       )
     },
     #' @description Get the process IDs of the local process and the