+++ tests/testthat/test-crew_controller_local.R
@@ -14,7 +14,7 @@ crew_test("crew_controller_local()", {
   expect_false(x$client$started)
   expect_false(x$started())
   expect_null(x$summary())
-  expect_false(x$autoscaling)
+  expect_null(x$loop)
   expect_equal(length(x$pids()), 1L)
   x$start()
   expect_true(x$empty())
@@ -35,8 +35,8 @@ crew_test("crew_controller_local()", {
     sort(
       c(
         "controller",
-        "tasks",
         "seconds",
+        "tasks",
         "success",
         "error",
         "crash",
@@ -45,20 +45,15 @@ crew_test("crew_controller_local()", {
       )
     )
   )
-  expect_equal(s$tasks, 0L)
   expect_true(x$client$started)
   expect_true(x$started())
   expect_equal(length(x$pids()), 2L)
   # first task
-  expect_equal(x$pushed, 0L)
-  expect_equal(x$popped, 0L)
   task <- x$push(
     command = Sys.getenv("CREW_WORKER"),
     name = "task"
   )
   expect_s3_class(task, "mirai")
-  expect_equal(x$pushed, 1L)
-  expect_equal(x$popped, 0L)
   expect_false(x$empty())
   expect_true(x$nonempty())
   x$wait(mode = "one", seconds_timeout = 5)
@@ -73,12 +68,9 @@ crew_test("crew_controller_local()", {
     seconds_interval = 0.5,
     seconds_timeout = 10
   )
-  expect_equal(x$pushed, 1L)
-  expect_equal(x$popped, 1L)
   out <- envir$out
   expect_true(x$empty())
   expect_false(x$nonempty())
-  expect_equal(x$summary()$tasks, 1L)
   expect_equal(x$summary()$error, 0L)
   expect_equal(x$summary()$warning, 0L)
   expect_equal(out$name, "task")
@@ -152,7 +144,6 @@ crew_test("crew_controller_local() substitute = FALSE and quick push", {
   expect_silent(x$validate())
   expect_false(x$client$started)
   x$start()
-  expect_equal(x$summary()$tasks, 0L)
   expect_equal(x$summary()$error, 0L)
   expect_equal(x$summary()$warning, 0L)
   command <- quote(sqrt(4L) + sqrt(9L))
@@ -247,7 +238,6 @@ crew_test("exit status and code", {
   task <- x$pop()
   expect_equal(task$status, "cancel")
   expect_equal(task$code, 20L)
-  expect_equal(x$client$resolved(), 3L)
 })
 
 crew_test("crew_controller_local() resource usage metric logging", {