+++ NEWS.md
@@ -1,10 +1,18 @@
-# crew 1.2.1.9002 (development)
+# crew 1.2.1.9003 (development)
 
 * Reinstate checks that were temporarily suppressed to help plugins transition to #217.
 * Support `reset_globals`, `reset_packages`, `reset_options`, and `garbage_collection` in `crew_controller_sequential()` (#217).
 * Change argument `tls` to `tlscert` in the call to `mirai::daemon()` (#227, @shikokuchuo).
 * Use `collections` queues instead of custom queues (#229).
 * Use `collections` dictionaries instead of hash environments to track tasks (#229).
+* Remove the long-deprecated `promise()` method of the controller.
+* Improve interoperability of `crew`'s workers by avoiding `nanonext::cv_value()` for condition variables (#225). Consequences:
+    * `wait(mode = "all")` no longer guarantees that a task is available for `pop()`. It just consumes a condition variable signal. `pop()` should always be checked for `NULL` return values.
+    * Until `mirai` gains a threaded dispatcher, `saturated()` needs to avoid the overhead of calling `status()` to get task counts. So the definition of "saturated" has changed: a controller is saturated if the number of *uncollected* tasks is greater than or equal to the maximum number of workers. Previously, it was the number of *unresolved* tasks.
+    * Internal counters `.pushed` and `.popped` are safely removed.
+    * Active bindings `pushed` and `popped` are removed.
+* Controllers no longer have a `unpopped()` method because it is unnecessary to export and it assumes only the controller submits tasks to the compute profile (affecting interoperability).
+* Allow a custom `later` loop in `autoscale()`.
 
 # crew 1.2.1